python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "code-researcher"
version = "1.0.0"
description = "Search GitHub for real code examples and inject them into LLM prompts"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["llm", "agents", "github", "code-search", "langchain", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
]
all = [
    "langchain-core>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
]

[project.scripts]
code-researcher = "cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/code-researcher"
Repository = "https://github.com/yourusername/code-researcher"
